Updated selection to pick students with the least number of available courses.
The idea here is when we are selecting students to pick, we'll use
the ones with the smaller number of available courses first, so
that they get higher choice. Easy for students to game the system,
but if the students are required to have equal number of choices,
we are less likely to have leftovers that weren't properly placed.